pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.9.1",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6",
    "distro==1.9.0",
    "dj-database-url==3.0.1",
    "django==5.2.5",
    "django-cors-headers==4.7.0",
    "django-sslserver==0.22",
    "djangorestframework==3.16.1",
    "drf-yasg==1.21.10",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "imagekitio==4.2.0",
    "inflection==0.5.1",
    "jiter==0.10.0",
    "openai==1.99.9",
    "packaging==25.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "waitress==3.0.2",
    "whitenoise==6.9.0",
]
